+++ src/r_things.h
@@ -60,6 +60,8 @@ void R_InitSpritesRes(void);
 
 void R_NearbySprites(void);
 
+vissprite_t *R_GetLastVisSprite(void); // [Nugget]
+
 #endif
 
 //----------------------------------------------------------------------------